+++ decidim-proposals/db/migrate/20180115155220_add_index_created_at_proposal_notes.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 class AddIndexCreatedAtProposalNotes < ActiveRecord::Migration[5.1]
   def change
     add_index :decidim_proposals_proposal_notes, :created_at