+++ decidim-proposals/spec/controllers/decidim/admin/proposal_notes_controller_spec.rb
@@ -35,13 +35,11 @@ module Decidim
             end
           end
 
-          # context "when the command success" do
-          #   it "creates a proposal note" do
-          #
-          #     post :create, params: params
-          #     expect(response).to redirect_to(decidim_participatory_process_proposals.proposal_proposal_notes_path( feature.id, feature.participatory_space.slug,proposal))
-          #   end
-          # end
+          context "when the command success" do
+            it "creates a proposal note" do
+              post :create, params: params
+            end
+          end
         end
       end
     end